Unify the types related to component providers
* Use the same pairing Provider – OptionalProvider everywhere
* Use the Provider types in the DashboardComponentsContext
* Move the types around to prevent cyclic dependencies

JIRA: RAIL-3905
