+++ libs/sdk-ui-dashboard/src/presentation/dashboardContexts/index.ts
@@ -6,3 +6,4 @@ export {
 } from "./DashboardCustomizationsContext";
 export { DashboardConfigProvider, useDashboardConfigContext } from "./DashboardConfigContext";
 export { ExportDialogContextProvider, useExportDialogContext } from "./ExportDialogContext";
+export * from "./types";
