+++ libs/sdk-ui-dashboard/src/presentation/index.ts
@@ -1,5 +1,7 @@
 // (C) 2021 GoodData Corporation
 export * from "./dashboard";
+// only export the types for this, not the actual code
+export * from "./dashboardContexts/types";
 export * from "./drill";
 export * from "./filterBar";
 export * from "./layout";
